[feat] : Adicionei nova regra de negócio para lidar com valores Doubles para inteiro, arredondando valores decimais, além de extrair para um método responsável por seguir essa regra.

+++ src/main/java/org/ondedoar/adapter/response/openstreet/RouteResponseDto.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class RouteResponseDto {
-    private Double duration;
-    private Double distance;
+    private Integer duration;
+    private Integer distance;
 
     public RouteResponseDto() {
     }
